Parameterise integration tests and other rename/refactoring

+++ auto-handbrake-cfr/src/main/java/com/willmolloy/handbrake/cfr/VideoArchiver.java
@@ -36,19 +36,18 @@ class VideoArchiver {
     try {
       log.debug("Archiving: {} -> {}", video.originalPath(), video.archivedPath());
 
-      if (Files.exists(video.archivedPath())) {
-        if (Files.mismatch(video.originalPath(), video.archivedPath()) == -1) {
-          log.warn("Archive file ({}) already exists", video.archivedPath());
-          if (!video.originalPath().equals(video.archivedPath())) {
-            log.info("Deleting: {}", video.originalPath());
-            Files.delete(video.originalPath());
-          }
-          return true;
-        } else {
+      if (!video.originalPath().equals(video.archivedPath())
+          && Files.exists(video.archivedPath())) {
+        if (Files.mismatch(video.originalPath(), video.archivedPath()) != -1) {
           log.error(
               "Archive file ({}) already exists but contents differ. Skipping archive process",
               video.archivedPath());
           return false;
+        } else {
+          log.warn("Archive file ({}) already exists", video.archivedPath());
+          log.info("Deleting: {}", video.originalPath());
+          Files.delete(video.originalPath());
+          return true;
         }
       }
